+++ changelog.md
@@ -1,1 +1,3 @@
-- Added config option for camera consumable item
+- Fixed bug that players in adventure mode can break the image frame
+- Fixed Image Frame not rendering in newer forge versions
+- Updated forge mappings